Refactor into separate files.
