+++ js/constants/pools.js
@@ -1,0 +1,17 @@
+const poolNames =
+  [ ["ItemRoom", "Item"]
+  , ["Shop", "Shop"]
+  , ["BossRoom", "Boss"]
+  , ["DevilRoom", "Devil"]
+  , ["AngelRoom", "Angel"]
+  , ["SecretRoom", "Secret"]
+  , ["Library", "Library"]
+  , ["GoldenChest", "Golden Chest"]
+  , ["RedChest", "Red Chest"]
+  , ["CurseRoom", "Curse"]
+  , ["Beggar", "Beggar"]
+  , ["DemonBeggar", "Demon Beggar"]
+  , ["KeyBeggar", "Key Beggar"]
+  ]
+
+export default poolNames
