+++ js/reducers/index.js
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux'
+import ranks from './ranks'
+
+const rootReducer = combineReducers(
+  { ranks }
+  )
+
+export default rootReducer
