+++ js/store/configureStore.js
@@ -1,0 +1,6 @@
+import {createStore} from 'redux'
+import rootReducer from '../reducers'
+
+export default function configureStore(initialState) {
+  return createStore(rootReducer, initialState)
+}
