+++ package.json
@@ -18,6 +18,7 @@
     "imports-loader": "^0.6.4",
     "react": "^0.13.3",
     "react-bootstrap": "^0.25.2",
+    "react-immutable-proptypes": "^1.2.1",
     "react-redux": "^2.1.1",
     "redux": "^2.0.0",
     "webpack": "^1.12.1",
